% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brmath.r
\name{brmath}
\alias{brmath}
\title{Base Functions for Analytical Solutions of Geohydrological Problems.}
\description{
Functions as defined in Part B ("Mathematical tools") in the book
"Analytical solutions of geohydrological problems" by G.A. Bruggeman.
}
\section{Error functions}{


\code{\link{P}}

\code{\link{Pconj}}

\code{\link{dPdx}}

\code{\link{dPdy}}

\code{\link{dPconjdx}}

\code{\link{dPconjdy}}

\code{\link{R}}

\code{\link{M}}
}

