% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{dPconjdx}
\alias{dPconjdx}
\title{Derivative of the Conjugate Polder function with respect to x, Chapter 3.1.2,
formula 31, page 863.}
\usage{
dPconjdx(x, y)
}
\arguments{
\item{x}{First argument in the Polder function}

\item{y}{Second argument in the Polder function}
}
\value{
Derivative of the Conjugate Polder function with respect to x
}
\description{
Derivative of the Conjugate Polder function with respect to x, Chapter 3.1.2,
formula 31, page 863.
}
\references{
Formula 31 page 863.
}
