% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{dPconjdy}
\alias{dPconjdy}
\title{Derivative of the Conjugate Polder function with respect to y, Chapter 3.1.2,
formula 32, page 863.}
\usage{
dPconjdy(x, y)
}
\arguments{
\item{x}{First argument in the Polder function}

\item{y}{Second argument in the Polder function}
}
\value{
Derivative of the Conjugate Polder function with respect to y
}
\description{
Derivative of the Conjugate Polder function with respect to y, Chapter 3.1.2,
formula 32, page 863.
}
\references{
Formula 32 page 863.
}
