% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{dPdx}
\alias{dPdx}
\title{Derivative of the Polder function with respect to x, Chapter 3.1.2,
formula 29, page 863.}
\usage{
dPdx(x, y)
}
\arguments{
\item{x}{First argument in the Polder function}

\item{y}{Second argument in the Polder function}
}
\value{
Derivative of the Polder function with respect to x
}
\description{
Derivative of the Polder function with respect to x, Chapter 3.1.2,
formula 29, page 863.
}
\references{
Formula 29 page 863.
}
