% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{dPdy}
\alias{dPdy}
\title{Derivative of the Polder function with respect to y, Chapter 3.1.2,
formula 30, page 863.}
\usage{
dPdy(x, y)
}
\arguments{
\item{x}{First argument in the Polder function}

\item{y}{Second argument in the Polder function}
}
\value{
Derivative of the Polder function with respect to y
}
\description{
Derivative of the Polder function with respect to y, Chapter 3.1.2,
formula 30, page 863.
}
\references{
Formula 30 page 863.
}
