% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{M}
\alias{M}
\title{M-function, Chapter 3.1.4, formula 49, page 869.}
\usage{
M(u, Alpha)
}
\arguments{
\item{u}{First argument in the M-function function}

\item{Alpha}{Second argument in the M-function}
}
\value{
Result of the M-function
}
\description{
M-function, Chapter 3.1.4, formula 49, page 869.
}
\references{
Formula 49 page 869.
}
