% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{P}
\alias{P}
\title{Polder function, Chapter 3.1.2, formula 24, page 862.}
\usage{
P(x, y)
}
\arguments{
\item{x}{First argument in the Polder function}

\item{y}{Second argument in the Polder function}
}
\value{
Result of the Polder function
}
\description{
Polder function, Chapter 3.1.2, formula 24, page 862.
}
\references{
Formula 24 page 862.
}
