% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{Pconj}
\alias{Pconj}
\title{Conjugate Polder function, Chapter 3.1.2, formula 25, page 862.}
\usage{
Pconj(x, y)
}
\arguments{
\item{x}{First argument in the Polder function}

\item{y}{Second argument in the Polder function}
}
\value{
Result of Conjugate the Polder function
}
\description{
Conjugate Polder function, Chapter 3.1.2, formula 25, page 862.
}
\references{
Formula 25 page 862.
}
