% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{R}
\alias{R}
\title{Resistance function, Chapter 3.1.3, formula 41, page 865.}
\usage{
R(x, y)
}
\arguments{
\item{x}{First argument in the Resistance function}

\item{y}{Second argument in the Resistance function}
}
\value{
Result of the Resistance function
}
\description{
Resistance function, Chapter 3.1.3, formula 41, page 865.
}
\references{
Formula 41 page 865.
}
